+++ src/applicationInternal/config/device.cpp
@@ -1,0 +1,1 @@
+#include "device.h"
