+++ src/applicationInternal/config/scene.cpp
@@ -1,0 +1,5 @@
+#include "scene.h"
+
+using namespace config;
+
+
