--- src/applicationInternal/devices/scene_registry.cpp
@@ -1,10 +1,0 @@
-#include "scene_registry.h"
-
-using namespace config;
-
-
-void CommandSequence::run() {
-    for(auto cmd: commands) {
-        cmd->execute();
-    }
-}
