--- src/applicationInternal/devices/scene_registry.h
@@ -1,18 +1,0 @@
-#pragma once
-#include <vector>
-#include <map>
-#include <unordered_map>
-#include <string>
-#include <stdint.h>
-#include <ArduinoJson.h>
-#include "deviceRegistry.h"
-
-
-namespace config {
-    struct CommandSequence {
-        std::vector<Command*> commands;
-        void run();
-    };
-
-
-}
