+++ src/guis/gui_devices.cpp
@@ -1,7 +1,7 @@
 #include "gui_devices.h"
 #include "applicationInternal/hardware/hardwarePresenter.h"
 #include "applicationInternal/gui/guiRegistry.h"
-#include "applicationInternal/devices/deviceRegistry.h"
+#include "applicationInternal/config/registry.h"
 #include "applicationInternal/commandHandler.h"
 #include "applicationInternal/omote_log.h"
